add skip support
This makes it possible for a file to be a forced-match.  That is,
even if the length/md5 are not available or don't actually match,
just accept whatever is on the remote server.

If the remote server does not have the file in question, then it
will still not be sent if skip is set to true.